requirements.txt
appdirs==1.4.3
get==0.0.0
packaging==16.8
post==0.0.0
public==0.0.0
pyparsing==2.2.0
query-string==0.0.0
request==0.0.0
setupfiles==0.0.0
six==1.10.0
